Added mpiprepsubband to the standard src tree. Modified the Makefile to handle it (a hack).
git-svn-id: file://localhost/data1/git/presto_svn/trunk@654 5fac236c-a0d8-0310-9745-ef116914eda1

+++ src/wapp.c
@@ -34,7 +34,7 @@ void get_WAPP_static(int *bytesperpt, int *bytesperblk, float *clip_sigma){
 }
 
 void set_WAPP_static(int ptsperblk, int bytesperpt, int bytesperblk, 
-		     int numchan, double clip_sigma, double dt){
+		     int numchan, float clip_sigma, double dt){
   ptsperblk_st = ptsperblk;
   bytesperpt_st = bytesperpt;
   bytesperblk_st = bytesperblk;
